+++ src/backend/api/index.ts
@@ -5,15 +5,15 @@ import rateLimit from "express-rate-limit";
 import RedisStore from "rate-limit-redis";
 import Redis from "ioredis";
 import BodyParser from "body-parser";
-import GitHubAPI from "../config/github-api";
-import { Installation, Subscription } from "../models";
-import verifyInstallation from "../jira/verify-installation";
-import logMiddleware from "../middleware/log-middleware";
-import JiraClient from "../models/jira-client";
-import uninstall from "../jira/uninstall";
+import GitHubAPI from "../../config/github-api";
+import { Installation, Subscription } from "../../models";
+import verifyInstallation from "../../jira/verify-installation";
+import logMiddleware from "../../middleware/log-middleware";
+import JiraClient from "../../models/jira-client";
+import uninstall from "../../jira/uninstall";
 import { serializeJiraInstallation, serializeSubscription } from "./serializers";
-import getRedisInfo from "../config/redis-info";
-import { elapsedTimeMetrics } from "../config/statsd";
+import getRedisInfo from "../../config/redis-info";
+import { elapsedTimeMetrics } from "../../config/statsd";
 
 const router = express.Router();
 const bodyParser = BodyParser.urlencoded({ extended: false });
