+++ src/backend/api/serializers.ts
@@ -1,6 +1,6 @@
-import JiraClient from "../models/jira-client";
-import Subscription from "../models/subscription";
-import Installation from "../models/installation";
+import JiraClient from "../../models/jira-client";
+import Subscription from "../../models/subscription";
+import Installation from "../../models/installation";
 import Logger from "bunyan";
 
 type SerializedSubscription = Pick<Subscription, "gitHubInstallationId" | "jiraHost" | "createdAt" | "updatedAt" | "syncStatus">;
